+++ src/main/java/de/shadowhunt/scm/subversion/v1_7/SubversionRepository1_7.java
@@ -92,7 +92,7 @@ public class SubversionRepository1_7 extends AbstractSubversionRepository<Subver
 	}
 
 	@Override
-	protected InputStream download0(final String sanatizedResource, final long version) {
+	protected InputStream download0(final String sanatizedResource, final int version) {
 		final URI uri;
 		if (version == HEAD_VERSION) {
 			uri = URI.create(repository + sanatizedResource);
@@ -106,7 +106,7 @@ public class SubversionRepository1_7 extends AbstractSubversionRepository<Subver
 	}
 
 	@Override
-	protected SubversionInfo info0(final String sanatizedResource, final long version, final boolean withCustomProperties) {
+	protected SubversionInfo info0(final String sanatizedResource, final int version, final boolean withCustomProperties) {
 		final URI uri;
 		if (version == HEAD_VERSION) {
 			uri = URI.create(repository + sanatizedResource);
